Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int a ,b ;
-    scanf("%d %d", &a , &b);
-    if(b ==  a*a)
+    double a ,b ;
+    scanf("%lf %lf", &a , &b);
+    if(b== a*a)
     printf("Yes");
     else
     printf("No");
